pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geoshield-mllm"
version = "0.1.0"
description = "Config-driven GeoShield-style geolocation privacy probing stack for multimodal APIs."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "fycorex", email = "zfysjtu24@sjtu.edu.cn" }]
dependencies = [
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "google-api-python-client>=2.120",
  "google-auth>=2.29",
  "google-auth-oauthlib>=1.2",
  "requests>=2.31",
]

[project.optional-dependencies]
providers = ["openai>=1.50", "anthropic>=0.34", "google-genai>=0.3", "dashscope>=1.20"]
dev = ["pytest>=8.0"]

[project.scripts]
geoshield-mllm = "geoshield_mllm.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

